extracted registration form to a component
